Orange is the colour between yellow and red on the spectrum of visible light. The human eyes perceive orange when observing light with a dominant wavelength between roughly 585 and 620 nanometres. In traditional colour theory, it is a secondary colour of pigments, produced by mixing yellow and red. In the RGB colour model, it is a tertiary colour. It is named after the fruit of the same name.
The orange colour of many fruits and vegetables, such as carrots, pumpkins, sweet potatoes, and oranges, comes from carotenes, a type of photosynthetic pigment. These pigments convert the light energy that the plants absorb from the Sun into chemical energy for the plants' growth. Similarly, the hues of autumn leaves are from the same pigment after chlorophyll is removed.

In Europe and the United States, surveys show that orange is the colour most associated with amusement, the unconventional, extroversion, warmth, fire, energy, activity, danger, taste and aroma, the autumn and Allhallowtide seasons, as well as having long been the national colour of the Netherlands and the House of Orange. It also serves as the political colour of the Christian democracy political ideology and most Christian democratic political parties.[4] In Asia, it is an important symbolic colour in Buddhism and Hinduism.[5]